+++ src/main/kotlin/org/ethereum/discv5/core/Router.kt
@@ -23,19 +23,19 @@ class Router {
      * TODO: real down nodes and network losses
      */
     fun route(from: Node, to: Enr, message: Message): List<Message> {
-        from.outgoingMessages.add(message.getSize())
+        from.outgoingMessages.add(message.getSize().toLong())
         from.roundtripLatency.add(listOf(Unit))
         val toNode = resolve(to)
         if (toNode == null) {
             logger.debug("Failed to route $message from Node${from.enr.toId()} to Node${to.toId()}, recipient not found")
             return emptyList()
         }
-        toNode.incomingMessages.add(message.getSize())
+        toNode.incomingMessages.add(message.getSize().toLong())
         return toNode.handle(message, from.enr).also { messages ->
             from.roundtripLatency.add(messages.map { Unit }.toList())
             messages.forEach {
-                toNode.outgoingMessages.add(it.getSize())
-                from.incomingMessages.add(it.getSize())
+                toNode.outgoingMessages.add(it.getSize().toLong())
+                from.incomingMessages.add(it.getSize().toLong())
             }
         }
     }
